feat: custom primary color and fix tooltip bug
