+++ packages/web3/src/connect-button/demos/simple.tsx
@@ -1,7 +1,7 @@
 import { ConnectButton } from '@ant-design/web3';
 
 const App: React.FC = () => {
-  return <ConnectButton />;
+  return <ConnectButton address="3ea2cfd153b8d8505097b81c87c11f5d05097c18" connected />;
 };
 
 export default App;
