+++ packages/web3/src/utils/format.ts
@@ -1,0 +1,4 @@
+export const fillWith0x = (address: string = ''): string => {
+  const filledAddress = address.startsWith('0x') ? address : `0x${address}`;
+  return filledAddress;
+};
