+++ packages/web3/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './browser';
+export * from './format';
